OrchestrationDriver/requirements.txt
azure-storage-file>=1.0
azure-storage-blob>=1.0
azure-common>=1.1.15
azure-mgmt-containerregistry>=1.0
azure-mgmt-resource>=1.1.0
azure-mgmt-containerinstance==1.1.0
azure-cosmos>=1.0.0
cassandra-driver>=1.0.0
pyyaml
opencv-python